requirements.txt
biopython==1.87
geneml==1.1.0
gffutils==0.13
goatools==1.6.4
h5py==3.15.1
helperlibs==0.2.2
ipykernel==6.30.1
keras==3.12.0
matplotlib==3.10.5
nbconvert==7.17.0
numpy==2.2.6
pandas==2.3.1
pyhmmer==0.12.0
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
taxoniq==1.0.3
tensorflow==2.20.0
